rename crest.js to updateCrest.js to improve readability
